+++ odoo/local-src/specific_zetes/models/stock_picking.py
@@ -3,7 +3,8 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 import random
 
-from odoo import models, fields, api
+from odoo import models, fields, api, _
+from odoo.exceptions import UserError
 
 from .. import constants
 
@@ -26,7 +27,14 @@ class StockPicking(models.Model):
         default=constants.AS_DEFAULT,
         required=True)
     is_zetes_error = fields.Boolean('Zetes error', default=False)
-    zetes_traceback = fields.Text('Zetes traceback')
+    zetes_picking_type = fields.Selection([
+        (constants.PICKING_ASSIGNMENT, 'Customer'),
+        (constants.PARKING_ASSIGNMENT, 'Parking'),
+        (constants.RESERVE_ASSIGNMENT, 'Reserve')],
+        string="Picking type",
+        default=constants.PICKING_ASSIGNMENT,
+        required=True
+    )
 
     @api.multi
     def assign_picking_checksum(self):
@@ -63,6 +71,38 @@ class StockPicking(models.Model):
             'operator_id': None,
         })
 
+    @api.multi
+    def validate_picking(self):
+        for picking in self:
+            # The method "do_new_transfer" is the method called when
+            # an user click on "Validate" on a picking.
+            result = picking.do_new_transfer()
+
+            # In Odoo this button will open a wizard in following case:
+            # 1. A wizard if no quantity has been defined on lines
+            #   (this wizard will set the quantity on each lines)
+            # 2. A wizard if we need to create a back order
+            if isinstance(result, dict):
+                model = result.get('res_model')
+                wizard = self.env[model].browse(int(result.get('res_id')))
+
+                # Fortunately these wizards have the same
+                # method "process" to execute the wizard
+                wizard.process()
+
+
+class PackOperationReserveRel(models.Model):
+    _name = 'pack.operation.reserve.rel'
+
+    pack_operation_id = fields.Many2one('stock.pack.operation',
+                                        string='Pack operation',
+                                        required=True)
+    reserve_location_id = fields.Many2one('stock.location',
+                                          string='Reserve',
+                                          required=True)
+    lot_id = fields.Many2one('stock.production.lot',
+                             string='Lot')
+
 
 class StockPackOperation(models.Model):
     _inherit = 'stock.pack.operation'
@@ -73,11 +113,184 @@ class StockPackOperation(models.Model):
         (constants.OP_SHORTPICKED, 'Shortpicked'),
         (constants.OP_SKIPPED, 'Skipped'),
         (constants.OP_CUT, 'Cut'),
-        (constants.OP_CANCELED, 'Canceled')
+        (constants.OP_CANCELED, 'Canceled / Full')
     ],
         string='Zetes state',
         default=constants.OP_DEFAULT,
         required=True)
+
+    @api.multi
+    def split_pack_op(self, new_qty, location_dest_id, lot_id=None):
+        """
+        Split the current pack operation in two pack operation.
+        One with the previous destination and a new one with the new
+        destination (location_dest_id) and the a specific quantity (new_qty).
+
+        1. Check the quantity (the new quantity cannot be greater than
+        the available quantity)
+        2. Create the new pack operation
+        3. (Optional) Reduce the quantity on the pack operation lot
+        4. Reduce the quantity on the pack operation
+        :param new_qty:
+        :param location_dest_id:
+        :param lot_id:
+        :return:
+        """
+        self.ensure_one()
+
+        # Step 1
+        quantity_available = self.product_qty - self.qty_done
+        if new_qty > quantity_available:
+            raise UserError(
+                _('You cannot split this pack operation because '
+                  'the new quantity (%s) is geater than '
+                  'the available quantity (%s)') %
+                (new_qty, quantity_available))
+
+        # Step 2
+        new_pack = self.copy({
+            'qty_done': 0.0,
+            'product_qty': new_qty,
+            'location_dest_id': location_dest_id,
+        })
+
+        # Step 3
+        if lot_id:
+            if not self.pack_lot_ids:
+                raise UserError(_('No pack operation found'))
+            pack_lot = self.pack_lot_ids.filtered(
+                lambda line: line.lot_id.id == lot_id)
+            if not pack_lot:
+                raise UserError(
+                    _('No pack operation found with ID %s' % lot_id))
+
+            lot_quantity_available = pack_lot.qty_todo - pack_lot.qty
+            if new_qty > lot_quantity_available:
+                raise UserError(
+                    _('You cannot split this pack operation lot because '
+                      'the new quantity (%s) is greater than '
+                      'the available quantity (%s)') %
+                    (new_qty, lot_quantity_available))
+
+            pack_lot.copy({
+                'operation_id': new_pack.id,
+                'qty_todo': new_qty,
+                'qty': 0
+            })
+
+            pack_lot.write({
+                'qty_todo': pack_lot.qty_todo - new_qty
+            })
+
+        # Step 4
+        self.write({
+            'product_qty': self.product_qty - new_qty
+        })
+
+        return new_pack
+
+    @api.multi
+    def add_qty(self, qty, lot_id=None):
+        """
+        Add a qty on the pack operation
+        :param qty: int - the qty to add
+        :param lot_id: int - the ID of the lot
+        :return:
+        """
+        self.ensure_one()
+
+        if not qty:
+            return
+
+        self.qty_done += qty
+
+        if not lot_id:
+            return
+
+        # When we have the lot, we will check if there no existing
+        # quantity for this lot.
+        pack_lot = \
+            self.pack_lot_ids.filtered(lambda line: line.lot_id.id == lot_id)
+
+        # If there no existing line (quantity) for this lot
+        # we will create a new line
+        if not len(pack_lot):
+            self.pack_lot_ids.create({
+                'operation_id': self.id,
+                'qty': qty,
+                'lot_id': lot_id,
+            })
+        # Otherwise we set the quantity for this lot
+        # We don't need to add the new quantity to the lot
+        # because Zetes send one request by lot
+        else:
+            pack_lot.write({'qty': qty})
+
+    @api.multi
+    def put_in_reserve(self, reserve_id):
+        """
+        Put the remaining quantity in the reserve
+        1. Split the pack operation to create the operation to the reserve
+        2. Reduce the quantity to do to the quantity done
+        3. (Optional) Reduce the quantity to do on the pack log to
+        the quantity done and remove empty pack operation lot
+        :param reserve_id:
+        :return:
+        """
+        # If no quantity has been taken, we simply change the destination
+        # of the current pack operation and set the quantity on the pack
+        # operation and pack operation lots
+        if not self.qty_done:
+            self.write({
+                'qty_done': self.product_qty,
+                'location_dest_id': reserve_id
+            })
+
+            for pack_lot in self.pack_lot_ids:
+                pack_lot.write({
+                    'qty': pack_lot.qty_todo,
+                })
+            return
+
+        quantity_remaining = self.product_qty - self.qty_done
+
+        # Step 1
+        new_pack = self.copy({
+            'qty_done': quantity_remaining,
+            'product_qty': quantity_remaining,
+            'location_dest_id': reserve_id
+        })
+
+        # Step 2
+        self.write({
+            'product_qty': self.qty_done,
+        })
+
+        # Step 3
+        pack_lot_to_unlink = self.env['stock.pack.operation.lot']
+        pack_lot_obj = self.env['stock.pack.operation.lot']
+        for pack_lot in self.pack_lot_ids:
+            if pack_lot.qty_todo == pack_lot.qty:
+                continue
+            qty = pack_lot.qty
+            qty_todo = pack_lot.qty_todo
+
+            qty_remaining = qty_todo - qty
+            pack_lot_obj.create({
+                'operation_id': new_pack.id,
+                'qty': qty_remaining,
+                'qty_todo': qty_remaining,
+                'lot_id': pack_lot.lot_id.id,
+            })
+
+            if not pack_lot.qty:
+                pack_lot_to_unlink |= pack_lot
+            else:
+                pack_lot.write({
+                    'qty_todo': pack_lot.qty
+                })
+
+        pack_lot_to_unlink.unlink()
 
 
 class StockPickingType(models.Model):